feat: config hot-reload, pill UI fixes, comparison update, onboard removal
- Config file watcher with ESM cache-busting (temp file copy) for live reload
- Dev server graceful shutdown (suppress ^C, clean exit message)
- Extracted dev utilities to src/utils/dev-utils.ts
- Sidebar version/language pill interaction fixes (mutual close, dropdown margins)
- Language pill responsive label (auto-compact when both pills present)
- Search client type fix for version badge rendering
- Icon parameter support docs for buttons, cards, callouts, collapsible
- titleAppend frontmatter option documented
- Comparison table updated with VitePress column across all READMEs
- Removed onboard script, merged --link flag into prep/verify
- Reordered package.json scripts for readability
- Updated CONTRIBUTING.md with new setup commands
