+++ README.hi.md
@@ -214,19 +214,20 @@ docmd plugin add <plugin-name>
 
 ## docmd क्यों?
 
-| विशेषता | docmd | Docusaurus | MkDocs | Mintlify |
-| :--------------- | :------------------------ | :------------------- | :-------------- | :--------------- |
-| **भाषा** | **Node.js** | React.js | Python | प्रोप्राइटरी |
-| **नेविगेशन** | **इंस्टैंट SPA** | React SPA | पेज रिलोड | होस्टेड SPA |
-| **आउटपुट** | **स्टैटिक HTML** | React Hydration | स्टैटिक HTML | होस्टेड |
-| **JS पेलोड** | **< 20kb** | > 200kb | न्यूनतम | मध्यम |
-| **वर्शनिंग** | **बिल्ट-इन** | फ़ाइल-आधारित (जटिल) | प्लगइन-आधारित | नेटिव |
-| **i18n** | **बिल्ट-इन** | नेटिव | थीम-आधारित | बीटा |
-| **सर्च** | **बिल्ट-इन (ऑफ़लाइन)** | Algolia (क्लाउड) | Lunr (बिल्ट-इन) | क्लाउड |
-| **PWA** | **बिल्ट-इन** | प्लगइन | कोई नहीं | होस्टेड |
-| **AI कॉन्टेक्स्ट** | **बिल्ट-इन (`llms.txt`)** | प्लगइन | कोई नहीं | प्रोप्राइटरी |
-| **सेटअप** | **तुरंत** | ~15 मिनट | ~10 मिनट | ~5 मिनट |
-| **लागत** | **मुफ़्त (OSS)** | मुफ़्त (OSS) | मुफ़्त (OSS) | फ्रीमियम |
+| विशेषता | docmd | Docusaurus | MkDocs Material | VitePress | Mintlify |
+| :--------------- | :------------------------ | :------------------- | :-------------- | :--------------- | :--------------- |
+| **भाषा** | **Node.js** | React.js | Python | Vue | SaaS |
+| **कॉन्फ़िग आवश्यक** | **कोई नहीं** | `docusaurus.config.js` | `mkdocs.yml` | `config.mts` | `mint.json` |
+| **प्रारंभिक पेलोड** | **~18kb** | ~250kb | ~40kb | ~50kb | ~120kb |
+| **नेविगेशन** | **इंस्टैंट SPA** | React SPA | पेज रिलोड | Vue SPA | होस्टेड SPA |
+| **वर्शनिंग** | **बिल्ट-इन** | नेटिव (जटिल) | mike प्लगइन | मैनुअल | नेटिव |
+| **i18n** | **बिल्ट-इन** | नेटिव (जटिल) | प्लगइन-आधारित | मैनुअल | नेटिव |
+| **सर्च** | **बिल्ट-इन (ऑफ़लाइन)** | Algolia (क्लाउड) | बिल्ट-इन | MiniSearch | क्लाउड |
+| **PWA** | **बिल्ट-इन** | कम्युनिटी प्लगइन | कोई नहीं | कोई नहीं | होस्टेड |
+| **AI कॉन्टेक्स्ट** | **बिल्ट-इन (`llms.txt`)** | मैनुअल | कोई नहीं | कोई नहीं | प्रोप्राइटरी |
+| **सेल्फ-होस्टेड** | **हाँ** | हाँ | हाँ | हाँ | नहीं |
+| **ज़ीरो-कॉन्फ़िग** | **`npx @docmd/core dev`** | नहीं | नहीं | नहीं | नहीं |
+| **लागत** | **मुफ़्त (OSS)** | मुफ़्त (OSS) | मुफ़्त (OSS) | मुफ़्त (OSS) | फ्रीमियम |
 
 सरलता से शुरू। बिना रुकावट के स्केल।
